Generated coefficients for y in latent variable model

Generated coefficients for y in latent variable model

Usage

makeCoefficients(
  p,
  k_unblocked,
  beta_low,
  beta_high,
  nblocks,
  sig_blocks,
  block_size
)

Arguments

p

Integer or numeric; the number of features that will be observed in x plus the number of latent variables (each corresponding to a cluster).

k_unblocked

Integer or numeric; the number of features in X that will have nonzero coefficients in the true model for y among those features not generated from the n_clusters latent variables (called "weak signal" features in the simulations from Faletto and Bien 2022). The coefficients on these features will be determined by beta_low.

beta_low

Integer or numeric; the maximum coefficient in the model for y among the k_unblocked features in X not generated from the latent variables. The coefficients of the features will be beta_low/sqrt(1:k_unblocked).

beta_high

Integer or numeric; the coefficient used for all sig_blocks latent variables that have nonzero coefficients in the true model for y.

nblocks

Integer or numeric; the number of latent variables that were generated, each of which will be associated with an observed cluster in X.

sig_blocks

Integer or numeric; the number of generated latent features that will have nonzero coefficients in the true model for y (all of them will have coefficient beta_latent). In particular, the first sig_blocks latent variables will have coefficient beta_latent, and the remaining nblocks

sig_blocks features will have coefficient 0. Must be less than or equal to n_clusters.

block_size

Integer or numeric; for each of the n_blocks latent variables, the covariance matrix will include the original latent feature plus block_size - 1 noisy proxies that are correlated with the latent variable.

Value

A named list with the following elements:

beta

A vector of length p containing the coefficients for the true model for y. All entries will equal 0 except for the sig_blocks latent variables that will have coefficient beta_high and the k_unblocked independent features with coefficient determined by beta_low.

blocked_dgp_vars

An integer vector of length sig_blocks containing the indices of the features corresponding to the latent features that will have nonzero coefficient beta_high in the true model for y.

sig_unblocked_vars

An integer vector of length k_unblocked containing the indices of the observed features that are independent of the blocked features and have coefficient beta_low in the true model for y. If k_unblocked = 0, this will just be NA.

insig_blocked_vars

An integer vector containing the indices of the features corresponding to the latent features that will have coefficient 0 in the true model for y. If nblocks=0, this will just be NA.

latent_vars

An integer vector of length nblocks containing the indices of all of the latent features.

References

Faletto, G., & Bien, J. (2022). Cluster Stability Selection. arXiv preprint arXiv:2201.00494. https://arxiv.org/abs/2201.00494.

Author

Gregory Faletto, Jacob Bien